+++ CHANGELOG.md
@@ -11,13 +11,17 @@ This project is [semver](https://semver.org/) based
 * Using tabs to simplify configuration
 * Added tab for information
 * Added toggle to store in the gallery
-* Is possible to see the runnning system config
+* Added toggle to insert a frame
+* Is possible to see the runnning system configuration
 * Is possible to go to the path where te images are stored
+* Added links to get more information
 
 ### Changed
 
 * The waiting time configuration was moved to an additional panel
 * Api key is masked once it has been used
+* Fixed error when trying to insert the image inside a frame
+* Fixed translation sync
 
 ## [0.8] - Gallery support